+++ CHANGELOG.md
@@ -5,12 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Color provider
 - Integration of VASM
 - Debugging with FS-UAE
 
 ### [0.5.0]
 ### Added
+- Color provider
 - Added documentation for CIAA and CIAB
 - Shows the values set to the registers
 - Show values in decimal / hexadecimal and binary